requirements.txt
bleach==6.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Login==0.6.2
Flask-Markdown==0.3
flask-paginate==2024.4.12
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==2.6.11
markdown2==2.4.13
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
webencodings==0.5.1
Werkzeug==2.2.3
zipp==3.15.0
